+++ CHANGELOG.md
@@ -1,14 +1,26 @@
+## 0.0.3
+
+* **BREAKING** Changed title's type of `ShadowBoxWithTitle` from _String_ to _Widget_
+* Exposed `themeMode` from `ThemeProvider`
+* Loader color now matches Theme's primary color
+* Added `itemOverviewIconBuilder` parameter to `DropdownSelector`
+* Changed `WillPopScope` widget to `PopScope` after deprecation
+
 ## 0.0.2+3
+
 * Changed loader asset
 * Added barrier color to loader dialog
 
 ## 0.0.2+2
+
 Fixed loader not hiding
 
 ## 0.0.2+1
+
 Fixed wrong default loader asset path
 
 ## 0.0.2
+
 * Added ExpandableText widget
 * Added Accordion widget
 * Added example showing widget through [Widgetbook](https://pub.dev/packages/widgetbook) package
@@ -18,7 +30,7 @@ Fixed wrong default loader asset path
 ## 0.0.1
 
 Initial release. Contains the following:
+
 * Light and Dark theme management
 * Loader management
 * Custom standard widgets [searchbar, loader, page, dropdown]
-